Add files via upload
To use my image recognition model (opening it will not work as it is encoded, directly give it the image) click on the howTouse.txt file, copy the code and run it. You MUST download ALL the files, or the code will not work. You can change the image on line 6. Thanks for using it!